\name{distMS2}
\alias{distMS2}
\title{
Distance metric for MS2 spectral similarity measurement
}
\description{
distance metric for MS2 spectral comparison. MS2 spectrum is provided as 2-col matrix.
}
\usage{
distMS2(a, b, topIon = 20)
}

\arguments{
  \item{a}{MS2 spectrum}
  \item{b}{MS2 spectrum}
  \item{topIon}{the number of the most intense ions used for comparison}
}
\details{
}
\value{

}
\references{
Cao M, Koulman A, Johnson LJ, Lane GA and Rasmussen S. 2008. Plant Physiology. Vol.146 No.4
}
\author{
Mingshu Cao
}
\note{
%%  ~~further notes~~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}

\keyword{ utilities }
\keyword{ cluster }
