\name{getMetaInfo}
\alias{getMetaInfo}
\title{
Get metadata information from data file in mzXML
}
\description{
print out some useful header information, such as instrumentation, ionization, range of mz and RT, and the number of MSn scans etc.  
}
\usage{
getMetaInfo(filename)
}

\arguments{
  \item{filename}{
  }
}
\details{
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mingshu Cao
}
\note{

}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
