\name{iontree-package}
\alias{iontree-package}
\alias{iontree}
\docType{package}
\title{
MSn-iontree: Ion tree management and analysis
}
\description{
management and analysis of ion fragmentation data
}
\details{
\tabular{ll}{
Package: \tab iontree\cr
Type: \tab Package\cr
LazyLoad: \tab yes\cr
}
This package provides functions to retrieve MSn fragmentation data, build MS2/MS3
ion trees from ion-trap low resoultion mass spectrometry and to create a relational 
database (SQLite-based) for routine management of ion tree data. Other functions 
include metrics for MS2 spectral simialrity measurement, iontree plotting and DB operations.
}
\author{
Mingshu Cao \cr
Maintainer: Mingshu Cao <mingshu.cao@agresearch.co.nz>
}
\references{
JRAP -- a Java library was used for parsing mzXML and mzML \cr
\url{http://sashimi.svn.sourceforge.net/viewvc/sashimi/trunk/jrap/stax/software/} \cr
}

\keyword{ package }
\seealso{
}
\examples{
}
