\name{plotSpectrum}
\alias{plotSpectrum}
\title{
Plot a spectrum
}
\description{
plot a mass spectrum
}
\usage{
plotSpectrum(x, y=NULL, top = 20, type = "h", scale100 = FALSE,
    digit.label = 1, col = "black", pos = 0, main = "",
    clickAddLabels = FALSE, ...)
}


\arguments{
  \item{x}{m/z}
  \item{y}{intensity}
  \item{top}{top intense m/z to be labelled}
  \item{type}{plot type}
  \item{scale100}{in scale of 0-100 if true}
  \item{digit.label}{m/z precision to be maintained}
  \item{col}{m/z label color}
  \item{pos}{m/z label position}
  \item{main}{title}
  \item{clickAddLabels}{click to add labels}
  \item{\dots}{as in plot}
}
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mingshu Cao
}
\note{
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}

\keyword{ aplot }
