\name{rs2iontree}
\alias{rs2iontree}
\title{
Convert a data frame into iontree
}
\description{
Convert resultset, a data frame retrieved from database into a list of iontree objects. 
}
\usage{
rs2iontree(rs)
}

\arguments{
  \item{rs}{resultset as a data frame}
}
\details{
}
\value{
a list of iontrees
}
\references{

}
\author{
Mingshu Cao
}
\note{
}

\seealso{
}

\examples{
}
\keyword{ utilities }
