\name{show-methods}
\docType{methods}
\alias{show-methods}
%%\alias{show,iontree-method}
%% remove many other alias --mcao
%%show is defined in iontree class, here is duplicated

\title{ Show methods for class iontree}
\description{
 show methods for class iontree
}
\section{Methods}{
\describe{
%%remove many other items

\item{\code{signature(object = "iontree")}}{show iontree object}

}}
\keyword{methods}

